+++ lib/carrierwave/base64/adapter.rb
@@ -1,4 +1,4 @@
-module Carrierwave
+module CarrierWave
   module Base64
     # Module with .mount_base64_uploader method, that is mixed in to
     # ActiveRecord::Base or Mongoid::Document::ClassMethods
@@ -7,7 +7,7 @@ module Carrierwave
       # string as input. It also accepts regular file uploads.
       #
       # @param attribute [Symbol] the attribute to mount this uploader on
-      # @param uploader_class [Carrierwave::Uploader] the uploader class to
+      # @param uploader_class [CarrierWave::Uploader] the uploader class to
       #   mount
       # @param options [Hash{Symbol => Object}] a set of options
       # @option options [Proc] :file_name Proc that must return a file name
@@ -22,9 +22,9 @@ module Carrierwave
         mount_uploader attribute, uploader_class, options
         options[:file_name] ||= proc { attribute }
 
-        Carrierwave::Base64::MountingHelper.check_for_deprecations(options)
+        CarrierWave::Base64::MountingHelper.check_for_deprecations(options)
 
-        Carrierwave::Base64::MountingHelper.define_writer(
+        CarrierWave::Base64::MountingHelper.define_writer(
           self, attribute, options
         )
       end