+++ lib/carrierwave/base64/railtie.rb
@@ -1,14 +1,14 @@
-module Carrierwave
+module CarrierWave
   module Base64
     # Railtie class to load the carrierwave-base64 adapters
     # Loads adapters for ActiveRecord and Mongoid
     class Railtie < Rails::Railtie
       ActiveSupport.on_load :active_record do
-        ActiveRecord::Base.extend Carrierwave::Base64::Adapter
+        ActiveRecord::Base.extend CarrierWave::Base64::Adapter
       end
 
       ActiveSupport.on_load :mongoid do
-        Mongoid::Document::ClassMethods.include Carrierwave::Base64::Adapter
+        Mongoid::Document::ClassMethods.include CarrierWave::Base64::Adapter
       end
     end
   end